Fix for a compilation warning at compiler build
/xmain-web/llvm/llvm/lib/Transforms/Intel_LoopTransforms/HIRArrayTranspose.cpp: In member function 'void {anonymous}::HIRArrayTranspose::MallocAnalyzer::processIntToPtrInst(llvm::loopopt::HLInst*, const llvm::IntToPtrInst*)':
/xmain-web/llvm/llvm/lib/Transforms/Intel_LoopTransforms/HIRArrayTranspose.cpp:353:69: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
       if (!PElemTy->isSized() || (PElemTy->getPrimitiveSizeInBits() !=
                                                                     ^
/xmain-web/llvm/llvm/lib/Transforms/Intel_LoopTransforms/HIRArrayTranspose.cpp: In member function 'bool {anonymous}::HIRArrayTranspose::MallocAnalyzer::isValidStridedUseRef(llvm::loopopt::RegDDRef*)':
/xmain-web/llvm/llvm/lib/Transforms/Intel_LoopTransforms/HIRArrayTranspose.cpp:659:73: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (!DestElemTy->isSized() || (DestElemTy->getPrimitiveSizeInBits() >
                                                                         ^

Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214322
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>

+++ llvm/lib/Transforms/Intel_LoopTransforms/HIRArrayTranspose.cpp
@@ -351,7 +351,7 @@ void HIRArrayTranspose::MallocAnalyzer::processIntToPtrInst(
       // %tFree = inttoptr.i64.elemTy*(%tMallocPtrToInt);
       // @free(&((i8*)(%tFree)[offset]));
       if (!PElemTy->isSized() || (PElemTy->getPrimitiveSizeInBits() !=
-                                  (HAT.ArrayElementSizeInBytes * 8))) {
+                                  uint64_t(HAT.ArrayElementSizeInBytes * 8))) {
         // Give up if destination type size does not match element size.
         HasValidMallocs = false;
         return;
@@ -657,7 +657,7 @@ bool HIRArrayTranspose::MallocAnalyzer::isValidStridedUseRef(RegDDRef *Ref) {
     // element size because consecutive elements before transpose are not
     // consecutive after transpose.
     if (!DestElemTy->isSized() || (DestElemTy->getPrimitiveSizeInBits() >
-                                   HAT.ArrayElementSizeInBytes * 8)) {
+                                   uint64_t(HAT.ArrayElementSizeInBytes * 8))) {
       return false;
     }
   }
